HydroBot/requirements.txt
discord.py
pandas
requests
python-dotenv
openai>=1.0.0